SQUASH: Rename andor files according to naming convention
