--- src/PlusDataCollection/Andor/vtkPlusAndorCamera.cxx
@@ -1,956 +1,0 @@
-/*=Plus=header=begin======================================================
-Program: Plus
-Copyright (c) Laboratory for Percutaneous Surgery. All rights reserved.
-See License.txt for details.
-=========================================================Plus=header=end*/
-
-#include "PlusConfigure.h"
-#include "vtkPlusDataSource.h"
-#include "vtkPlusAndorVideoSource.h"
-#include "ATMCD32D.h"
-#include "igtlOSUtil.h" // for Sleep
-#include "opencv2/imgproc.hpp"
-#include "opencv2/imgcodecs.hpp"
-
-
-vtkStandardNewMacro(vtkPlusAndorVideoSource);
-
-// put these here so there is no public dependence on OpenCV
-cv::Mat cvCameraIntrinsics;
-cv::Mat cvDistanceCoefficients;
-cv::Mat cvFlatCorrection;
-cv::Mat cvBiasCorrection;
-
-// ----------------------------------------------------------------------------
-void vtkPlusAndorVideoSource::PrintSelf(ostream& os, vtkIndent indent)
-{
-  this->Superclass::PrintSelf(os, indent);
-
-  os << indent << "Shutter: " << Shutter << std::endl;
-  os << indent << "ExposureTime: " << ExposureTime << std::endl;
-  os << indent << "Binning: " << HorizontalBins << " " << VerticalBins << std::endl;
-  os << indent << "HSSpeed: " << HSSpeed[0] << HSSpeed[1] << std::endl;
-  os << indent << "VSSpeed: " << VSSpeed << std::endl;
-  os << indent << "PreAmpGain: " << PreAmpGain << std::endl;
-  os << indent << "AcquisitionMode: " << AcquisitionMode << std::endl;
-  os << indent << "ReadMode: " << ReadMode << std::endl;
-  os << indent << "TriggerMode: " << TriggerMode << std::endl;
-  os << indent << "UseCooling: " << UseCooling << std::endl;
-  os << indent << "CoolTemperature: " << CoolTemperature << std::endl;
-  os << indent << "SafeTemperature: " << SafeTemperature << std::endl;
-  os << indent << "CurrentTemperature: " << CurrentTemperature << std::endl;
-  os << indent << "CameraIntrinsics: " << cvCameraIntrinsics << std::endl;
-  os << indent << "DistanceCoefficients: " << cvDistanceCoefficients << std::endl;
-  os << indent << "UseFrameCorrections: " << UseFrameCorrections << std::endl;
-  os << indent << "FlatCorrection: " << flatCorrection << std::endl;
-  os << indent << "BiasCorrection: " << biasCorrection << std::endl;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::ReadConfiguration(vtkXMLDataElement* rootConfigElement)
-{
-  LOG_TRACE("vtkPlusAndorVideoSource::ReadConfiguration");
-  XML_FIND_DEVICE_ELEMENT_REQUIRED_FOR_READING(deviceConfig, rootConfigElement);
-
-  // Must initialize the system before setting parameters
-  checkStatus(Initialize(""), "Initialize");
-
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, Shutter, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(float, ExposureTime, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, PreAmpGain, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, AcquisitionMode, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, ReadMode, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, TriggerMode, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, CoolTemperature, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, SafeTemperature, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, VSSpeed, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, HorizontalBins, deviceConfig);
-  XML_READ_SCALAR_ATTRIBUTE_OPTIONAL(int, VerticalBins, deviceConfig);
-
-  XML_READ_BOOL_ATTRIBUTE_OPTIONAL(UseCooling, deviceConfig);
-  XML_READ_BOOL_ATTRIBUTE_OPTIONAL(UseFrameCorrections, deviceConfig);
-
-  deviceConfig->GetVectorAttribute("HSSpeed", 2, HSSpeed);
-  deviceConfig->GetVectorAttribute("CameraIntrinsics", 9, cameraIntrinsics);
-  deviceConfig->GetVectorAttribute("DistanceCoefficients", 4, distanceCoefficients);
-  flatCorrection = deviceConfig->GetAttribute("FlatCorrection");
-  biasCorrection = deviceConfig->GetAttribute("BiasCorrection");
-
-  cvCameraIntrinsics = cv::Mat(3, 3, CV_64FC1, cameraIntrinsics);
-  cvDistanceCoefficients = cv::Mat(1, 4, CV_64FC1, distanceCoefficients);
-  this->SetFlatCorrectionImage(flatCorrection); // load and normalize if needed
-  this->SetBiasCorrectionImage(biasCorrection); // load the image
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::WriteConfiguration(vtkXMLDataElement* rootConfigElement)
-{
-  XML_FIND_DEVICE_ELEMENT_REQUIRED_FOR_WRITING(deviceConfig, rootConfigElement);
-
-  deviceConfig->SetIntAttribute("Shutter", this->Shutter);
-  deviceConfig->SetFloatAttribute("ExposureTime", this->ExposureTime);
-  deviceConfig->SetIntAttribute("PreAmpGain", this->PreAmpGain);
-  deviceConfig->SetIntAttribute("AcquisitionMode", this->AcquisitionMode);
-  deviceConfig->SetIntAttribute("ReadMode", this->ReadMode);
-  deviceConfig->SetIntAttribute("TriggerMode", this->TriggerMode);
-  deviceConfig->SetIntAttribute("CoolTemperature", this->CoolTemperature);
-  deviceConfig->SetIntAttribute("SafeTemperature", this->SafeTemperature);
-  deviceConfig->SetIntAttribute("VSSpeed", this->VSSpeed);
-  deviceConfig->SetIntAttribute("HorizontalBins", this->HorizontalBins);
-  deviceConfig->SetIntAttribute("VerticalBins", this->VerticalBins);
-
-  deviceConfig->SetVectorAttribute("HSSpeed", 2, HSSpeed);
-  deviceConfig->SetVectorAttribute("CameraIntrinsics", 9, cameraIntrinsics);
-  deviceConfig->SetVectorAttribute("DistanceCoefficients", 4, distanceCoefficients);
-  deviceConfig->SetAttribute("FlatCorrection", flatCorrection.c_str());
-  deviceConfig->SetAttribute("BiasCorrection", biasCorrection.c_str());
-
-  XML_WRITE_BOOL_ATTRIBUTE(UseCooling, deviceConfig);
-  XML_WRITE_BOOL_ATTRIBUTE(UseFrameCorrections, deviceConfig);
-
-  return PLUS_SUCCESS;
-}
-
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::NotifyConfigured()
-{
-  if(this->OutputChannels.empty())
-  {
-    LOG_ERROR("No output channels defined for vtkPlusAndorVideoSource. Cannot proceed.");
-    this->CorrectlyConfigured = false;
-    return PLUS_FAIL;
-  }
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-std::string vtkPlusAndorVideoSource::GetSdkVersion()
-{
-  std::ostringstream versionString;
-
-  char SDKVersion[256];
-  checkStatus(GetVersionInfo(AT_SDKVersion, SDKVersion, sizeof(SDKVersion)), "GetVersionInfo");
-  versionString << "Andor SDK version: "  << SDKVersion << std::ends;
-
-  return versionString.str();
-}
-
-
-//----------------------------------------------------------------------------
-vtkPlusAndorVideoSource::vtkPlusAndorVideoSource()
-{
-  this->RequirePortNameInDeviceSetConfiguration = true;
-
-  this->StartThreadForInternalUpdates = false; // frames should not be acquired automatically
-  this->AcquisitionRate = 1.0; // this controls the frequency
-}
-
-// ----------------------------------------------------------------------------
-vtkPlusAndorVideoSource::~vtkPlusAndorVideoSource()
-{
-  if(!this->Connected)
-  {
-    this->Disconnect();
-  }
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::InitializeAndorCamera()
-{
-  checkStatus(Initialize(""), "Initialize");
-
-  // Check the safe temperature, and the maximum allowable temperature on the camera.
-  // Use the min of the two as the safe temp.
-  int MinTemp, MaxTemp;
-  unsigned result = checkStatus(GetTemperatureRange(&MinTemp, &MaxTemp), "GetTemperatureRange");
-  if(result == DRV_SUCCESS)
-  {
-    LOG_INFO("The temperature range for the connected Andor Camera is: " << MinTemp << " and " << MaxTemp);
-  }
-
-  if(MaxTemp < this->SafeTemperature)
-  {
-    this->SafeTemperature = MaxTemp;
-  }
-  if(this->CoolTemperature < MinTemp || this->CoolTemperature > MaxTemp)
-  {
-    LOG_ERROR("Requested temperature for Andor camera is out of range");
-    return PLUS_FAIL;
-  }
-
-  int x, y;
-  checkStatus(GetDetector(&x, &y), "GetDetector");
-  frameSize[0] = static_cast<unsigned>(x);
-  frameSize[1] = static_cast<unsigned>(y);
-
-  // init to binning of 1 (meaning no binning), and full sensor size
-  checkStatus(SetImage(this->HorizontalBins, this->VerticalBins, 1, x, 1, y), "SetImage");
-
-  checkStatus(PrepareAcquisition(), "PrepareAcquisition");
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-void vtkPlusAndorVideoSource::InitializePort(DataSourceArray& port)
-{
-  for(unsigned i = 0; i < port.size(); i++)
-  {
-    port[i]->SetPixelType(VTK_UNSIGNED_SHORT);
-    port[i]->SetImageType(US_IMG_BRIGHTNESS);
-    port[i]->SetOutputImageOrientation(US_IMG_ORIENT_MF);
-    port[i]->SetInputImageOrientation(US_IMG_ORIENT_MF);
-    port[i]->SetInputFrameSize(frameSize);
-
-    LOG_INFO("Andor source initialized. ID: " << port[i]->GetId());
-  }
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::InternalConnect()
-{
-  LOG_TRACE("vtkPlusAndorVideoSource::InternalConnect");
-  if(this->InitializeAndorCamera() != PLUS_SUCCESS)
-  {
-    LOG_ERROR("Andor camera failed to initialize.");
-    return PLUS_FAIL;
-  }
-
-  this->GetVideoSourcesByPortName("BLIraw", BLIraw);
-  this->GetVideoSourcesByPortName("BLIrectified", BLIrectified);
-  this->GetVideoSourcesByPortName("GrayRaw", GrayRaw);
-  this->GetVideoSourcesByPortName("GrayRectified", GrayRectified);
-
-  if(BLIraw.size() + BLIrectified.size() + BLIdark.size() + GrayRaw.size() + GrayRectified.size() + GrayDark.size() == 0)
-  {
-    vtkPlusDataSource* aSource = nullptr;
-    if(this->GetFirstActiveOutputVideoSource(aSource) != PLUS_SUCCESS || aSource == nullptr)
-    {
-      LOG_ERROR("Standard data sources are not defined, and unable to retrieve the video source in the capturing device.");
-      return PLUS_FAIL;
-    }
-    BLIraw.push_back(aSource); // this is the default port
-  }
-
-  this->InitializePort(BLIraw);
-  this->InitializePort(BLIrectified);
-  this->InitializePort(GrayRaw);
-  this->InitializePort(GrayRectified);
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::InternalDisconnect()
-{
-  LOG_DEBUG("Disconnecting from Andor");
-  if(IsRecording())
-  {
-    this->InternalStopRecording();
-  }
-
-  int status;
-  checkStatus(::IsCoolerOn(&status), "IsCoolerOn");
-
-  if(status && this->CoolerMode == 0)
-  {
-    LOG_INFO("CoolerMode 0 and Cooler is still ON. Turning off the cooler and waiting for warmup. Do not unplug the camera from power.");
-    WaitForWarmup();
-  }
-
-  checkStatus(FreeInternalMemory(), "FreeInternalMemory");
-
-  unsigned result = checkStatus(ShutDown(), "ShutDown");
-  if(result == DRV_SUCCESS)
-  {
-    LOG_INFO("Andor camera shut down successfully.");
-  }
-
-  return PLUS_SUCCESS;
-}
-
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::InternalStartRecording()
-{
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::InternalStopRecording()
-{
-  return PLUS_SUCCESS;
-}
-
-
-// ----------------------------------------------------------------------------
-float vtkPlusAndorVideoSource::GetCurrentTemperature()
-{
-  checkStatus(GetTemperatureF(&this->CurrentTemperature), "GetTemperatureF");
-  return this->CurrentTemperature;
-}
-
-// ----------------------------------------------------------------------------
-void vtkPlusAndorVideoSource::WaitForCooldown()
-{
-  if(this->UseCooling == false)
-  {
-    return;
-  }
-  int status;
-  checkStatus(::IsCoolerOn(&status), "IsCoolerOn");
-  if(!status)
-  {
-    TurnCoolerON();
-  }
-  while(checkStatus(GetTemperatureF(&this->CurrentTemperature), "GetTemperatureF") != DRV_TEMPERATURE_STABILIZED)
-  {
-    igtl::Sleep(5000); // wait a bit
-  }
-}
-
-// ----------------------------------------------------------------------------
-void vtkPlusAndorVideoSource::WaitForWarmup()
-{
-  if(this->UseCooling == false)
-  {
-    return;
-  }
-  int status;
-  checkStatus(::IsCoolerOn(&status), "IsCoolerOn");
-  if(status)
-  {
-    TurnCoolerOFF();
-  }
-  GetCurrentTemperature(); // updates this->CurrentTemperature
-  if(this->CurrentTemperature < this->SafeTemperature)
-  {
-    while(this->CurrentTemperature < this->SafeTemperature)
-    {
-      igtl::Sleep(5000); // wait a bit
-      GetCurrentTemperature(); // logs the status and temperature
-    }
-  }
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::AcquireFrame(float exposure, int shutterMode, int binning, int vsSpeed, int hsSpeed)
-{
-  unsigned rawFrameSize = frameSize[0] * frameSize[1];
-  rawFrame.resize(rawFrameSize, 0);
-
-  checkStatus(::SetExposureTime(exposure), "SetExposureTime");
-  checkStatus(::SetShutter(1, shutterMode, 0, 0), "SetShutter");
-
-  int hbin = binning > 0 ? binning : this->HorizontalBins;
-  int vbin = binning > 0 ? binning : this->VerticalBins;
-  checkStatus(::SetImage(hbin, vbin, 1, 1024, 1, 1024), "Binning");
-
-  int vsInd = vsSpeed >= 0 ? vsSpeed : this->VSSpeed;
-  checkStatus(::SetVSSpeed(vsInd), "SetVSSpeed");
-
-  int hsInd = hsSpeed >= 0 ? hsSpeed : this->HSSpeed[1];
-  checkStatus(::SetHSSpeed(this->HSSpeed[0], hsInd), "SetHSSpeed");
-
-  checkStatus(StartAcquisition(), "StartAcquisition");
-  unsigned result = checkStatus(WaitForAcquisition(), "WaitForAcquisition");
-  if(result == DRV_NO_NEW_DATA)   // Log a more specific log message for WaitForAcquisition
-  {
-    LOG_ERROR("Non-Acquisition Event occurred.(e.g. CancelWait() called)");
-  }
-  this->currentTime = vtkIGSIOAccurateTimer::GetSystemTime();
-
-  // iKon-M 934 has 16-bit digitization
-  // https://andor.oxinst.com/assets/uploads/products/andor/documents/andor-ikon-m-934-specifications.pdf
-  // so we choose 16-bit unsigned
-  // GetMostRecentImage() is 32 bit signed variant
-  checkStatus(GetMostRecentImage16(&rawFrame[0], rawFrameSize), "GetMostRecentImage16");
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-void vtkPlusAndorVideoSource::AddFrameToDataSource(DataSourceArray& ds)
-{
-  for(unsigned i = 0; i < ds.size(); i++)
-  {
-    if(ds[i]->AddItem(&rawFrame[0],
-                      US_IMG_ORIENT_MF,
-                      frameSize, VTK_UNSIGNED_SHORT,
-                      1, US_IMG_BRIGHTNESS, 0,
-                      this->FrameNumber,
-                      currentTime,
-                      UNDEFINED_TIMESTAMP,
-                      nullptr) != PLUS_SUCCESS)
-    {
-      LOG_WARNING("Error adding item to AndorCamera video source " << ds[i]->GetSourceId());
-    }
-    else
-    {
-      LOG_INFO("Success adding item to AndorCamera video source " << ds[i]->GetSourceId());
-    }
-  }
-}
-
-
-// ----------------------------------------------------------------------------
-void vtkPlusAndorVideoSource::ApplyFrameCorrections()
-{
-  cv::Mat cvIMG(frameSize[0], frameSize[1], CV_16UC1, &rawFrame[0]); // uses rawFrame as buffer
-  cv::Mat floatImage;
-  cvIMG.convertTo(floatImage, CV_32FC1);
-  cv::Mat result;
-
-  cv::subtract(floatImage, cvBiasCorrection, floatImage, cv::noArray(), CV_32FC1);
-  LOG_INFO("Applied constant bias correction");
-
-  // Divide the image by the 32-bit floating point correction image
-  cv::divide(floatImage, cvFlatCorrection, floatImage, 1, CV_32FC1);
-  LOG_INFO("Applied multiplicative flat correction");
-
-  // OpenCV's lens distortion correction
-  cv::undistort(floatImage, result, cvCameraIntrinsics, cvDistanceCoefficients);
-  LOG_INFO("Applied lens distortion correction");
-  result.convertTo(cvIMG, CV_16UC1);
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::AcquireBLIFrame(int binning, int vsSpeed, int hsSpeed, float exposureTime)
-{
-  WaitForCooldown();
-  AcquireFrame(exposureTime, 0, binning, vsSpeed, hsSpeed);
-  ++this->FrameNumber;
-  AddFrameToDataSource(BLIraw);
-
-  if(this->UseFrameCorrections)
-  {
-    ApplyFrameCorrections();
-    AddFrameToDataSource(BLIrectified);
-  }
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::AcquireGrayscaleFrame(int binning, int vsSpeed, int hsSpeed, float exposureTime)
-{
-  WaitForCooldown();
-  AcquireFrame(exposureTime, 0, binning, vsSpeed, hsSpeed);
-  ++this->FrameNumber;
-  AddFrameToDataSource(GrayRaw);
-
-  if(this->UseFrameCorrections)
-  {
-    ApplyFrameCorrections();
-    AddFrameToDataSource(GrayRectified);
-  }
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::AcquireBiasFrame(std::string biasFilePath, int binning, int vsSpeed, int hsSpeed)
-{
-  AcquireFrame(0, 2, binning, vsSpeed, hsSpeed);
-  ++this->FrameNumber;
-  cv::Mat saveImage(frameSize[0], frameSize[1], CV_16UC1, &rawFrame[0]);
-  cv::imwrite(biasFilePath, saveImage);
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetBiasCorrectionImage(std::string biasFilePath)
-{
-  try
-  {
-    cvBiasCorrection = cv::imread(biasFilePath, cv::IMREAD_GRAYSCALE);
-  }
-  catch(...)
-  {
-    LOG_ERROR("Could not load bias correction image from file: " << biasFilePath);
-    return PLUS_FAIL;
-  }
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetFlatCorrectionImage(std::string flatFilePath)
-{
-  try
-  {
-    cvFlatCorrection = cv::imread(flatFilePath, cv::IMREAD_GRAYSCALE);
-    double maxVal = 0.0;
-    cv::minMaxLoc(cvFlatCorrection, nullptr, &maxVal);
-    if(maxVal > 1.0)   // we need to normalize the image to [0.0, 1.0] range
-    {
-      cv::Mat temp;
-      cvFlatCorrection.convertTo(temp, CV_32FC1, 1.0 / maxVal);
-      cvFlatCorrection = temp;
-    }
-  }
-  catch(...)
-  {
-    LOG_ERROR("Could not load flat correction image from file: " << flatFilePath);
-    return PLUS_FAIL;
-  }
-}
-
-// Setup the Andor camera parameters ----------------------------------------------
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetShutter(int shutter)
-{
-  this->Shutter = shutter;
-  checkStatus(::SetShutter(1, this->Shutter, 0, 0), "SetShutter");
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-int vtkPlusAndorVideoSource::GetShutter()
-{
-  return this->Shutter;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetExposureTime(float exposureTime)
-{
-  this->ExposureTime = exposureTime;
-  checkStatus(::SetExposureTime(this->ExposureTime), "SetExposureTime");
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-float vtkPlusAndorVideoSource::GetExposureTime()
-{
-  return this->ExposureTime;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetHorizontalBins(int bins)
-{
-  int x, y;
-  checkStatus(GetDetector(&x, &y), "GetDetector");  // full sensor size
-  unsigned status = checkStatus(::SetImage(bins, this->VerticalBins, 1, x, 1, y), "SetImage");
-  if(status != DRV_SUCCESS)
-  {
-    LOG_ERROR("SetImage failed while changing horizontal bins.");
-    return PLUS_FAIL;
-  }
-  this->HorizontalBins = bins;
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetVerticalBins(int bins)
-{
-  int x, y;
-  checkStatus(GetDetector(&x, &y), "GetDetector");  // full sensor size
-  unsigned status = checkStatus(::SetImage(this->HorizontalBins, bins, 1, x, 1, y), "SetImage");
-  if(status != DRV_SUCCESS)
-  {
-    LOG_ERROR("SetImage failed while changing vertical bins.");
-    return PLUS_FAIL;
-  }
-  this->VerticalBins = bins;
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetHSSpeed(int type, int index)
-{
-  unsigned status = checkStatus(::SetHSSpeed(type, index), "SetHSSpeed");
-  if(status != DRV_SUCCESS)
-  {
-    LOG_ERROR("SetHSSpeed command failed.");
-    return PLUS_FAIL;
-  }
-  HSSpeed[0] = type;
-  HSSpeed[1] = index;
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetVSSpeed(int index)
-{
-  unsigned status = checkStatus(::SetVSSpeed(index), "SetVSSpeed");
-  if(status != DRV_SUCCESS)
-  {
-    LOG_ERROR("SetVSSpeed command failed.");
-    return PLUS_FAIL;
-  }
-  this->VSSpeed = index;
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetPreAmpGain(int preAmpGain)
-{
-  this->PreAmpGain = preAmpGain;
-  unsigned status = checkStatus(::SetPreAmpGain(this->PreAmpGain), "SetPreAmpGain");
-  if(status == DRV_P1INVALID)
-  {
-    LOG_ERROR("Minimum threshold outside valid range (1-65535).");
-  }
-  else if(status == DRV_P2INVALID)
-  {
-    LOG_ERROR("Maximum threshold outside valid range.");
-  }
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-int vtkPlusAndorVideoSource::GetPreAmpGain()
-{
-  return this->PreAmpGain;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetAcquisitionMode(int acquisitionMode)
-{
-  this->AcquisitionMode = acquisitionMode;
-  checkStatus(::SetAcquisitionMode(this->AcquisitionMode), "SetAcquisitionMode");
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-int vtkPlusAndorVideoSource::GetAcquisitionMode()
-{
-  return this->AcquisitionMode;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetReadMode(int readMode)
-{
-  this->ReadMode = readMode;
-  checkStatus(::SetReadMode(this->ReadMode), "SetReadMode");
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-int vtkPlusAndorVideoSource::GetReadMode()
-{
-  return this->ReadMode;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetTriggerMode(int triggerMode)
-{
-  this->TriggerMode = triggerMode;
-  checkStatus(::SetTriggerMode(this->TriggerMode), "SetTriggerMode");
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-int vtkPlusAndorVideoSource::GetTriggerMode()
-{
-  return this->TriggerMode;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetUseFrameCorrections(bool useFrameCorrections)
-{
-  this->UseFrameCorrections = useFrameCorrections;
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-bool vtkPlusAndorVideoSource::GetUseFrameCorrections()
-{
-  return this->UseFrameCorrections;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetUseCooling(bool useCooling)
-{
-  int coolerStatus = 1;
-  unsigned result = checkStatus(::IsCoolerOn(&coolerStatus), "IsCoolerOn");
-  this->UseCooling = useCooling;
-  if(useCooling && coolerStatus == 0)
-  {
-    TurnCoolerON();
-  }
-  else if(useCooling == false && coolerStatus == 1)
-  {
-    TurnCoolerOFF();
-  }
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-bool vtkPlusAndorVideoSource::GetUseCooling()
-{
-  return this->UseCooling;
-}
-
-// ----------------------------------------------------------------------------
-int vtkPlusAndorVideoSource::IsCoolerOn()
-{
-  int coolerStatus = 1;
-  unsigned result = checkStatus(::IsCoolerOn(&coolerStatus), "IsCoolerOn");
-  if(result == DRV_SUCCESS)
-  {
-    return coolerStatus;
-  }
-  return -1;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::TurnCoolerON()
-{
-  unsigned result = checkStatus(CoolerON(), "CoolerON");
-  if(result == DRV_SUCCESS)
-  {
-    LOG_INFO("Temperature controller switched ON.");
-    checkStatus(SetTemperature(this->CoolTemperature), "SetTemperature");
-    return PLUS_SUCCESS;
-  }
-  LOG_ERROR("CoolerON command failed to execute.");
-  return PLUS_FAIL;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::TurnCoolerOFF()
-{
-  unsigned result = checkStatus(CoolerOFF(), "CoolerOFF");
-  if(result == DRV_SUCCESS)
-  {
-    LOG_INFO("Temperature controller switched OFF.");
-    return PLUS_SUCCESS;
-  }
-  LOG_ERROR("CoolerOFF command failed to execute.");
-  return PLUS_FAIL;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetCoolerMode(int mode)
-{
-  unsigned result = checkStatus(::SetCoolerMode(mode), "SetCoolerMode");
-  if(result == DRV_SUCCESS)
-  {
-    this->CoolerMode = mode;
-    if(mode == 1)
-    {
-      LOG_INFO("Cooler mode set to 1. Temperature will be maintained on ShutDown.");
-    }
-    else
-    {
-      LOG_INFO("Cooler mode set to 0. Camera will return to ambient temperature on ShutDown.");
-    }
-    return PLUS_SUCCESS;
-  }
-  LOG_ERROR("SetCoolerMode command failed to execute.");
-  return PLUS_FAIL;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetCoolTemperature(int coolTemp)
-{
-  this->CoolTemperature = coolTemp;
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-int vtkPlusAndorVideoSource::GetCoolTemperature()
-{
-  return this->CoolTemperature;
-}
-
-// ----------------------------------------------------------------------------
-PlusStatus vtkPlusAndorVideoSource::SetSafeTemperature(int safeTemp)
-{
-  this->SafeTemperature = safeTemp;
-
-  return PLUS_SUCCESS;
-}
-
-// ----------------------------------------------------------------------------
-int vtkPlusAndorVideoSource::GetSafeTemperature()
-{
-  return this->SafeTemperature;
-}
-
-// ----------------------------------------------------------------------------
-unsigned int vtkPlusAndorVideoSource::checkStatus(unsigned int returnStatus, std::string functionName)
-{
-  if(returnStatus == DRV_SUCCESS)
-  {
-    return returnStatus;
-  }
-  else if(returnStatus == DRV_NOT_INITIALIZED)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Driver is not initialized.");
-  }
-  else if(returnStatus == DRV_ACQUIRING)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Not allowed. Currently acquiring data.");
-  }
-  else if(returnStatus == DRV_P1INVALID)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Parameter 1 not valid.");
-  }
-  else if(returnStatus == DRV_P2INVALID)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Parameter 2 not valid.");
-  }
-  else if(returnStatus == DRV_P3INVALID)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Parameter 3 not valid.");
-  }
-  else if(returnStatus == DRV_P4INVALID)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Parameter 4 not valid.");
-  }
-  else if(returnStatus == DRV_P5INVALID)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Parameter 5 not valid.");
-  }
-  else if(returnStatus == DRV_P6INVALID)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Parameter 6 not valid.");
-  }
-  else if(returnStatus == DRV_P7INVALID)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Parameter 7 not valid.");
-  }
-  else if(returnStatus == DRV_ERROR_ACK)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Unable to communicate with card.");
-  }
-  else if(returnStatus == DRV_TEMP_OFF)
-  {
-    LOG_INFO("Cooler is OFF. Current temperature is " << this->CurrentTemperature << " °C");
-  }
-  else if(returnStatus == DRV_TEMPERATURE_STABILIZED)
-  {
-    LOG_INFO("Temperature has stabilized at " << this->CurrentTemperature << " °C");
-  }
-  else if(returnStatus == DRV_TEMPERATURE_NOT_REACHED)
-  {
-    LOG_INFO("Cooling down, current temperature is " << this->CurrentTemperature << " °C");
-  }
-  else if(returnStatus == DRV_TEMP_DRIFT)
-  {
-    LOG_INFO("Temperature had stabilised but has since drifted. Current temperature is " << this->CurrentTemperature << " °C");
-  }
-  else if(returnStatus == DRV_TEMP_NOT_STABILIZED)
-  {
-    LOG_INFO("Temperature reached but not stabilized. Current temperature is " << this->CurrentTemperature << " °C");
-  }
-  else if(returnStatus == DRV_VXDNOTINSTALLED)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; VxD not loaded.");
-  }
-  else if(returnStatus == DRV_INIERROR)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Unable to load DETECTOR.INI.");
-  }
-  else if(returnStatus == DRV_COFERROR)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Unable to load *.COF.");
-  }
-  else if(returnStatus == DRV_FLEXERROR)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Unable to load *.RBF.");
-  }
-  else if(returnStatus == DRV_ERROR_FILELOAD)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Unable to load *.COF or *.RBF files.");
-  }
-  else if(returnStatus == DRV_USBERROR)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Unable to detect USB device or not USB 2.0.");
-  }
-  else if(returnStatus == DRV_ERROR_NOCAMERA)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; No camera found.");
-  }
-  else if(returnStatus == DRV_GENERAL_ERRORS)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; An error occured while obtaining the number of available cameras.");
-  }
-  else if(returnStatus == DRV_INVALID_MODE)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Invalid mode or mode not available.");
-  }
-  else if(returnStatus == DRV_ERROR_PAGELOCK)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Unable to allocate memory.");
-  }
-  else if(returnStatus == DRV_INVALID_FILTER)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Filter not available for current acquisition.");
-  }
-  else if(returnStatus == DRV_BINNING_ERROR)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Range not a multiple of horizontal binning.");
-  }
-  else if(returnStatus == DRV_SPOOLSETUPERROR)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Error with spool settings.");
-  }
-  else if(returnStatus == DRV_IDLE)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; The system is not currently acquiring.");
-  }
-  else if(returnStatus == DRV_NO_NEW_DATA)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; There is no new data yet.");
-  }
-  else if(returnStatus == DRV_ERROR_CODES)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Problem communicating with camera.");
-  }
-  else if(returnStatus == DRV_LOAD_FIRMWARE_ERROR)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Error loading firmware.");
-  }
-  else if(returnStatus == DRV_NOT_SUPPORTED)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Feature not supported.");
-  }
-  else if(returnStatus == DRV_RANDOM_TRACK_ERROR)
-  {
-    LOG_ERROR("Failed AndorSDK operation: " << functionName            \
-              << "; Invalid combination of tracks.");
-  }
-  else
-  {
-    LOG_WARNING("Possible failed AndorSDK operation: " << functionName            \
-                << "; Unknown return code " << returnStatus << "returned.");
-  }
-
-  return returnStatus;
-}
